+++ packages/components/editor/src/components/extensions/image/index.tsx
@@ -1,0 +1,39 @@
+import Image from "@tiptap/extension-image";
+
+export const ExtendedImage = Image.extend({
+  addAttributes() {
+    return {
+      src: {
+        default: undefined,
+      },
+      alt: {
+        default: undefined,
+      },
+      placeholder: {
+        default: undefined,
+      },
+      title: {
+        default: undefined,
+      },
+      width: {
+        default: undefined,
+      },
+      height: {
+        default: undefined,
+      },
+      style: {
+        default: "",
+      },
+    };
+  },
+  parseHTML() {
+    return [
+      {
+        tag: "img",
+      },
+    ];
+  },
+  renderHTML({ HTMLAttributes }) {
+    return ["img", HTMLAttributes];
+  },
+});