+++ packages/components/editor/src/components/menu-item.tsx
@@ -7,23 +7,26 @@ export default function ToolbarItem({
   title,
   onClick,
   isActive = () => false,
+  isDisabled,
   ...props
 }: {
   icon: IconTypes;
   title?: string;
   onClick?: () => void;
   isActive?: () => boolean;
+  isDisabled?: () => boolean;
 }) {
   return (
     <Button
       isIconOnly
-      type="button"
       onClick={onClick}
       title={title}
       aria-label={title}
       size="sm"
       variant={isActive() ? "solid" : "light"}
+      disabled={isDisabled?.()}
       {...props}
+      type="button"
     >
       <EditorIcon name={icon} />
     </Button>