+++ packages/components/editor/src/editor.tsx
@@ -16,10 +16,6 @@ export const Editor = (props: EditorProps) => {
     editor,
     helperText,
   } = useEditor(props);
-
-  const json = editor?.getJSON()
-  console.log("json:->", json)
-
   return (
     <Component {...getBaseProps()}>
       <div {...getEditorProps()}>