+++ packages/components/file-upload/src/single/use-file-upload.tsx
@@ -20,7 +20,7 @@ import {
 type Props = Omit<UploadVariants, "isDragActive"> & {
   value: string;
   onValueChange: (value: string) => void;
-  onFileSelect: (file: File) => (void | string) | Promise<void | string>;
+  onFileSelect: (file: File) => void | Promise<void>;
   className?: string;
   classNames?: SlotsToClasses<UploadSlots>;
   onError?: (error: FileError) => void;